fix build, move interval tree
