+++ ydb/core/tx/columnshard/engines/reader/common_reader/constructor/read_metadata.h
@@ -14,25 +14,6 @@ class TLockSharingInfo;
 
 namespace NKikimr::NOlap::NReader::NCommon {
 
-struct TPortionIntervalTreeValueTraits: NRangeTreap::TDefaultValueTraits<std::shared_ptr<TPortionInfo>> {
-    struct THashFnc {
-        ui64 operator()(const std::shared_ptr<TPortionInfo>& value) const {
-            return THash<TPortionAddress>()(value->GetAddress());
-        }
-    };
-
-    static bool Less(const std::shared_ptr<TPortionInfo>& a, const std::shared_ptr<TPortionInfo>& b) noexcept {
-        return a->GetAddress() == b->GetAddress();
-    }
-
-    static bool Equal(const std::shared_ptr<TPortionInfo>& a, const std::shared_ptr<TPortionInfo>& b) noexcept {
-        return a->GetAddress() == b->GetAddress();
-    }
-};
-
-using TPortionIntervalTree =
-    NRangeTreap::TRangeTreap<NArrow::TSimpleRow, std::shared_ptr<TPortionInfo>, NArrow::TSimpleRow, TPortionIntervalTreeValueTraits>;
-
 class TSpecialReadContext;
 class IDataSource;
 class ISourcesConstructor {
